[ty] Move diagnostics API for the server
